Boss Lane Attack Feature Testing
This is a checkpoint for Nathan to pickup and keep working from

+++ Assets/AAA_FallenFate/Scripts/Nathan'sBossScripts/BossStringController.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class BossStringController : MonoBehaviour
 {
     //bool alreadyAttacked;
-    public int ShadowUptime = 2;
-    public int DelayBetweenAttacks = 2;
+    public float ShadowUptime = 2;
+    public float DelayBetweenAttacks = 2;
 
     public GameObject String1;
     public GameObject String2;
